requirements.txt
discord.py
mcstatus
mcipc
python-dotenv
